requirements.txt
crewai
crewai-tools
langchain-openai
python-dotenv
beautifulsoup4
requests
selenium
webdriver-manager
langgraph